Create 5hmCtweet.sh
This small example script receives the output of rss-filter.pl or biorxiv-filter.pl and sends it to Twitter. The file specified by the -keyf argument to `ttytter` determines which account is used (so you need to substitude your own keyfile for $KEYFILE above). A date-stamped log is saved of everything tweeted.